Revert "Feat/o auth page"
