+++ client/app/app.factory.js
@@ -149,7 +149,7 @@
         unshareDish(data.dish_id, data.user_id);
       });
 
-      socketFactory.on('billsSentToGuests', function(data) { //why is this socket.on?
+      socket.on('billsSentToGuests', function(data) {
         console.log("Heard 'billsSentToGuests' in appFactory.data:", data);
         services.data.billData = data;
         $rootScope.$broadcast('billsSentToGuests');
