--- client/app/oAuth/oAuth.controller.js
@@ -1,17 +1,0 @@
-(function() {
-  'use strict';
-
-  angular.module('Piecemeal')
-  .controller('OAuthCtrl', OAuthCtrl);
-
-  OAuthCtrl.$inject = ['oAuthFactory'];
-
-  function OAuthCtrl(oAuthFactory) {
-    var self = this;
-    self.venmoAuth = function () {
-      oAuthFactory.venmoLogin();
-    }
-    
-  }
-
-})();
