--- client/app/oAuth/oAuth.html
@@ -1,6 +1,0 @@
-<div id="oAuth" ng-cloak>
-  <button ng-click="oAuth.venmoAuth()">Create Bill With Venmo</button>
-  <div>Or</div>
-  <button>Create Event</button>
-  <input type="text" placeholder="Enter your name here">
-</div>
