+++ client/app/socket.module.js
@@ -25,6 +25,7 @@
     function init() {
       var ioRoom = $window.location.origin + '/' + $window.sessionStorage.code;
       $window.socket = io(ioRoom);
+      // console.log('Joining ioRoom: ', ioRoom);
     }
 
     function on(eventName, callback) {
