+++ client/index.html
@@ -43,8 +43,6 @@
     <script src="app/app.module.js"></script>
     <script src="app/app.routes.js"></script>
     <script src="app/socket.module.js"></script>
-    <script src="app/oAuth/oAuth.controller.js"></script>
-    <script src="app/oAuth/oAuth.factory.js"></script>    
     <script src="app/addDish/addDish.controller.js"></script>
     <script src="app/addDish/addDish.factory.js"></script>
     <script src="app/allDishes/allDishes.controller.js"></script>
