+++ package.json
@@ -26,7 +26,6 @@
     "bower": "^1.6.5",
     "cookie-parser": "^1.4.0",
     "express": "^4.13.3",
-    "express-cors": "0.0.3",
     "knex": "^0.9.0",
     "lodash": "^3.10.1",
     "matchdep": "^1.0.0",
@@ -49,6 +48,7 @@
     "grunt-contrib-connect": "^0.11.2",
     "grunt-contrib-sass": "^0.9.2",
     "grunt-contrib-uglify": "^0.11.0",
+    "grunt-injector": "^0.6.1",
     "grunt-open": "^0.2.3",
     "grunt-wiredep": "^2.0.0",
     "jasmine-core": "^2.4.1",
